+++ packages/lib/src/main.ts
@@ -1,0 +1,3 @@
+export * from '@collectium/core'
+export type * from '@collectium/core'
+