+++ packages/preset/github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @collectium/preset-github
 
+## 0.1.2
+
+### Patch Changes
+
+- add collectium general package
+
+- Updated dependencies []:
+  - @collectium/core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes